+++ bindings/imgui_bundle/demos_cpp/demos_immapp/demo_implot_markdown.cpp
@@ -9,7 +9,7 @@
 
 int main(int, char**)
 {
-    // This call is specific to the ImGui Bundle interactive manual. In a standard application, you could write:
+    // This call is specific to the ImGui Bundle Explorer. In a standard application, you could write:
     //         HelloImGui::SetAssetsFolder("my_assets"); // (By default, HelloImGui will search inside "assets")
     ChdirBesideAssetsFolder();